# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Forge 软件发布管理平台 API
  description: 本文档提供了 Forge 软件发布管理平台的详细 API 文档，包括认证、软件管理、统计分析等各个模块的接口说明。
  version: 1.0.0
  contact:
    name: Forge 开发团队
    email: dev@forge.com
servers:
  - url: http://localhost:5000/api
    description: 开发服务器
  - url: https://api.forge.com/api
    description: 生产服务器
security:
  - BearerAuth: []
paths:
...
  /software:
    get:
      tags:
        - 软件管理
      summary: 获取软件空间列表
      description: 获取所有软件空间的列表
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取软件空间列表成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 获取软件空间列表成功
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SoftwareSpace'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - 软件管理
      summary: 创建软件空间
      description: 创建一个新的软件空间
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: 软件名称
                  example: Forge
                description:
                  type: string
                  description: 软件描述
                  example: 一个强大的软件发布管理平台
                author:
                  type: string
                  description: 作者
                  example: Forge Team
                webhook_url:
                  type: string
                  description: Webhook URL
                  example: https://example.com/webhook
                  format: uri
      responses:
        '201':
          description: 软件空间创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 软件空间创建成功
                  data:
                    $ref: '#/components/schemas/SoftwareSpace'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /software/{space_id}:
    get:
      tags:
        - 软件管理
      summary: 获取软件空间详情
      description: 获取指定软件空间的详细信息
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SpaceId'
      responses:
        '200':
          description: 获取软件空间详情成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 获取软件空间详情成功
                  data:
                    $ref: '#/components/schemas/SoftwareSpace'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - 软件管理
      summary: 更新软件空间
      description: 更新指定软件空间的信息
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SpaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 软件名称
                  example: Forge
                description:
                  type: string
                  description: 软件描述
                  example: 一个强大的软件发布管理平台
                author:
                  type: string
                  description: 作者
                  example: Forge Team
                webhook_url:
                  type: string
                  description: Webhook URL
                  example: https://example.com/webhook
                  format: uri
      responses:
        '200':
          description: 软件空间更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 软件空间更新成功
                  data:
                    $ref: '#/components/schemas/SoftwareSpace'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - 软件管理
      summary: 删除软件空间
      description: 删除指定的软件空间及其所有相关数据
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SpaceId'
      responses:
        '200':
          description: 软件空间删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /software/{space_id}/regenerate-api-key:
    post:
      tags:
        - 软件管理
      summary: 重新生成API密钥
      description: 为指定软件空间重新生成API密钥
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SpaceId'
      responses:
        '200':
          description: API密钥重新生成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: API密钥重新生成成功
                  data:
                    type: object
                    properties:
                      api_key:
                        type: string
                        description: 新的API密钥
                        example: a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0u1v2w3x4y5z6
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /software/{space_id}/versions:
    get:
      tags:
        - 软件管理
      summary: 获取软件版本列表
      description: 获取指定软件空间的所有版本
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SpaceId'
      responses:
        '200':
          description: 获取软件版本列表成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 获取软件版本列表成功
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SoftwareVersion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - 软件管理
      summary: 创建软件版本
      description: 为指定软件空间创建一个新版本
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SpaceId'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - version
              properties:
                file:
                  type: string
                  format: binary
                  description: 软件文件
                version:
                  type: string
                  description: 版本号
                  example: 1.0.0
                release_note:
                  type: string
                  description: 发布说明
                  example: 初始版本发布
                documentation_url:
                  type: string
                  description: 文档链接
                  example: https://example.com/docs
                  format: uri
      responses:
        '201':
          description: 软件版本创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 软件版本创建成功
                  data:
                    $ref: '#/components/schemas/SoftwareVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /versions/{version_id}:
    get:
      tags:
        - 软件管理
      summary: 获取软件版本详情
      description: 获取指定软件版本的详细信息
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200':
          description: 获取软件版本详情成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 获取软件版本详情成功
                  data:
                    $ref: '#/components/schemas/SoftwareVersion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - 软件管理
      summary: 更新软件版本
      description: 更新指定软件版本的信息
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/VersionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                release_note:
                  type: string
                  description: 发布说明
                  example: 修复了一些bug
                documentation_url:
                  type: string
                  description: 文档链接
                  example: https://example.com/docs
                  format: uri
      responses:
        '200':
          description: 软件版本更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 软件版本更新成功
                  data:
                    $ref: '#/components/schemas/SoftwareVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - 软件管理
      summary: 删除软件版本
      description: 删除指定的软件版本及其文件
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200':
          description: 软件版本删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /versions/{version_id}/publish:
    put:
      tags:
        - 软件管理
      summary: 发布/下架软件版本
      description: 发布或下架指定的软件版本
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/VersionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - publish
              properties:
                publish:
                  type: boolean
                  description: 是否发布
                  example: true
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 软件版本发布成功
                  data:
                    $ref: '#/components/schemas/SoftwareVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /download/{version_id}:
    get:
      tags:
        - 软件管理
      summary: 下载软件版本（管理员）
      description: 下载指定的软件版本（仅限管理员）
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200':
          description: 下载成功
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # 公开API接口
  /public/{api_key}:
    get:
      tags:
        - 公开API
      summary: 获取软件空间信息（公开API）
      description: 通过API密钥获取软件空间的基本信息
      security: []  # 无需认证
      parameters:
        - $ref: '#/components/parameters/ApiKey'
      responses:
        '200':
          description: 获取软件空间信息成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 获取软件空间信息成功
                  data:
                    $ref: '#/components/schemas/SoftwareSpace'
        '404':
          $ref: '#/components/responses/NotFound'
  /public/{api_key}/versions:
    get:
      tags:
        - 公开API
      summary: 获取软件版本列表（公开API）
      description: 通过API密钥获取软件空间的所有已发布版本
      security: []  # 无需认证
      parameters:
        - $ref: '#/components/parameters/ApiKey'
      responses:
        '200':
          description: 获取软件版本列表成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 获取软件版本列表成功
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SoftwareVersion'
        '404':
          $ref: '#/components/responses/NotFound'
  /public/{api_key}/download/{version}:
    get:
      tags:
        - 公开API
      summary: 下载软件版本（公开API）
      description: 通过API密钥下载指定版本的软件
      security: []  # 无需认证
      parameters:
        - $ref: '#/components/parameters/ApiKey'
        - name: version
          in: path
          required: true
          description: 版本号或"latest"表示最新版本
          schema:
            type: string
            example: 1.0.0
      responses:
        '200':
          description: 下载成功
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'

  # 统计分析相关接口
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token认证
  schemas:
...
    SoftwareSpace:
      type: object
      properties:
        id:
          type: integer
          description: 软件空间ID
          example: 1
        name:
          type: string
          description: 软件名称
          example: Forge
        description:
          type: string
          description: 软件描述
          example: 一个强大的软件发布管理平台
        author:
          type: string
          description: 作者
          example: Forge Team
        webhook_url:
          type: string
          description: Webhook URL
          example: https://example.com/webhook
          format: uri
        created_by:
          type: integer
          description: 创建者ID
          example: 1
        created_at:
          type: string
          format: date-time
          description: 创建时间
          example: 2023-01-01T00:00:00Z
        updated_at:
          type: string
          format: date-time
          description: 更新时间
          example: 2023-01-01T00:00:00Z
        versions_count:
          type: integer
          description: 版本数量
          example: 5
        downloads_count:
          type: integer
          description: 下载次数
          example: 1000
        api_key:
          type: string
          description: API密钥（仅部分接口返回）
          example: a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0u1v2w3x4y5z6
    SoftwareVersion:
      type: object
      properties:
        id:
          type: integer
          description: 版本ID
          example: 1
        space_id:
          type: integer
          description: 软件空间ID
          example: 1
        version:
          type: string
          description: 版本号
          example: 1.0.0
        file_path:
          type: string
          description: 文件路径（仅部分接口返回）
          example: /storage/software/1/v1.0.0_file.exe
        file_size:
          type: integer
          description: 文件大小（字节）
          example: 1048576
        file_hash:
          type: string
          description: 文件哈希值
          example: a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0u1v2w3x4y5z6
        file_size_human:
          type: string
          description: 人类可读的文件大小
          example: 1.00MB
        release_note:
          type: string
          description: 发布说明
          example: 初始版本发布
        documentation_url:
          type: string
          description: 文档链接
          example: https://example.com/docs
          format: uri
        is_published:
          type: boolean
          description: 是否已发布
          example: true
        publish_date:
          type: string
          format: date-time
          description: 发布时间
          example: 2023-01-01T00:00:00Z
        created_by:
          type: integer
          description: 创建者ID
          example: 1
        created_at:
          type: string
          format: date-time
          description: 创建时间
          example: 2023-01-01T00:00:00Z
        updated_at:
          type: string
          format: date-time
          description: 更新时间
          example: 2023-01-01T00:00:00Z
        download_count:
          type: integer
          description: 下载次数
          example: 100
...
  parameters:
    SpaceId:
      name: space_id
      in: path
      required: true
      description: 软件空间ID
      schema:
        type: integer
    VersionId:
      name: version_id
      in: path
      required: true
      description: 版本ID
      schema:
        type: integer
    ApiKey:
      name: api_key
      in: path
      required: true
      description: API密钥
      schema:
        type: string
  responses:
    Success:
      description: 操作成功
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              message:
                type: string
                example: 操作成功
              data:
                type: object
                nullable: true
                example: null
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 请求参数错误
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: INVALID_PARAMETER
                  message:
                    type: string
                    example: 请求参数无效
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 未授权，请登录
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: UNAUTHORIZED
                  message:
                    type: string
                    example: 未授权访问
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 禁止访问
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: FORBIDDEN
                  message:
                    type: string
                    example: 权限不足
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 资源不存在
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: RESOURCE_NOT_FOUND
                  message:
                    type: string
                    example: 资源不存在
...